requirements.txt
numpy
torch
torchvision
tqdm
networkx
pillow
jupyter
ipython
nbconvert
tifffile
pyyaml
matplotlib